Issue 21 - WIP - Add speedometer score plots + refactoring
Replace speedometer graphs from using a ratio comparing Firefox and Chrome
to the absolute scores.

We're refactoring the backend to use query parameters and use modeIds
to filter through the data returned by AWFY. This guarantees that if
the order of an array changes on AWFY we won't be grabbing data based on
the index of an array.