Merge Beta data with Nightly and Canary graph
First of all, we don't have a Beta release criteria, thus,
the graph comparing it against Canary was meaningless.

Second, we can draw the Beta line as a dark line with Nightly and Canary
and not cloud the main purpose of the graph.

Third, we can now clearly compare Nightly against Beta.