Initial coverage support
Add coverage script, use `nyc` with `--all` to include all files that are not included by the tests,
use test/**/*.test.js to reach deep into the test directory. Run nyc report after the
fact to show a summary. Use `lcov` as the reporter to create both the text version
and the HTML version of the code coverage report.

Use the `include` option to only consider files under the `src/` directory.
Use the `extension` option to include the .jsx files which are excluded by default.