Drop Dashboard and globally change background-color and color
The original index.css set the background-color to black and the font color
to white. This changes it and uses the standard default colors for each propery.

All React components with in-styling were inverting these to the default values.
Thus, changing the global values requires adjusting the various React components.

This change is in preparation to incorporate the JS performance dashboard (arewefastyet.com).