Some refactor & reference to Netlify
In order to have access to the Open Source tier with Netlify we need
to make reference to them. It could be move to an about page if we like
but it looks nice at the home page with the black background.

This change also makes some improvements to the DashboardPage component
by allowing overwriting the CSS properties for the root div.